# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Currículo por Vaga com IA
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface CreateSessionBody {
  careerGoal?: string | null;
  applicationVolume?: string | null;
}

export type UpdateSessionBodyStructuredProfileJson = {
  [key: string]: unknown;
} | null;

export interface UpdateSessionBody {
  careerGoal?: string | null;
  applicationVolume?: string | null;
  interviewRate?: string | null;
  templateChoice?: string | null;
  fullName?: string | null;
  email?: string | null;
  phone?: string | null;
  location?: string | null;
  linkedinUrl?: string | null;
  portfolioUrl?: string | null;
  summary?: string | null;
  profileSourceType?: string | null;
  structuredProfileJson?: UpdateSessionBodyStructuredProfileJson;
  jobTitle?: string | null;
  company?: string | null;
  jobDescription?: string | null;
}

export interface ImportLinkedinBody {
  linkedinUrl: string;
}

export interface Experience {
  title: string;
  company: string;
  period?: string | null;
  description?: string | null;
}

export interface Education {
  degree: string;
  institution: string;
  period?: string | null;
}

export interface Project {
  name: string;
  description?: string | null;
  url?: string | null;
}

export interface StructuredProfile {
  fullName?: string | null;
  currentRole?: string | null;
  email?: string | null;
  phone?: string | null;
  location?: string | null;
  linkedinUrl?: string | null;
  portfolioUrl?: string | null;
  summary?: string | null;
  experiences?: Experience[];
  skills?: string[];
  education?: Education[];
  certifications?: string[];
  languages?: string[];
  projects?: Project[];
}

export interface ImportResult {
  success: boolean;
  profile?: StructuredProfile | null;
  error?: string | null;
}

export interface Session {
  id: string;
  status: string;
  careerGoal?: string | null;
  applicationVolume?: string | null;
  interviewRate?: string | null;
  templateChoice?: string | null;
  fullName?: string | null;
  email?: string | null;
  phone?: string | null;
  location?: string | null;
  linkedinUrl?: string | null;
  portfolioUrl?: string | null;
  summary?: string | null;
  profileSourceType?: string | null;
  structuredProfileJson?: unknown | null;
  jobTitle?: string | null;
  company?: string | null;
  jobDescription?: string | null;
  atsScore?: number | null;
  jobAnalysisJson?: unknown | null;
  previewResumeJson?: unknown | null;
  finalResumeJson?: unknown | null;
  paymentStatus?: string | null;
  generationStatus?: string | null;
  stripeCheckoutSessionId?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface GenerationStatusResponse {
  generationStatus: string;
  atsScore?: number | null;
  previewResumeJson?: unknown | null;
  jobAnalysisJson?: unknown | null;
}

export interface CheckoutResponse {
  checkoutUrl: string;
}

export type StripeWebhookBody = { [key: string]: unknown };

export type StripeWebhook200 = {
  received: boolean;
};