# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Currículo por Vaga com IA
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sessions/{sessionId}/checkout:
    post:
      operationId: createCheckout
      tags: [sessions]
      summary: Create Stripe checkout session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Checkout URL returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
...
  /stripe/webhook:
    post:
      operationId: stripeWebhook
      tags: [stripe]
      summary: Handle Stripe webhook events
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Webhook handled
          content:
            application/json:
              schema:
                type: object
                properties:
                  received:
                    type: boolean
                required:
                  - received
components:
  schemas:
...
    CheckoutResponse:
      type: object
      properties:
        checkoutUrl:
          type: string
      required:
        - checkoutUrl